+++ src/Algebra.cc
@@ -13,7 +13,7 @@
 using bsm::LorentzVector;
 using bsm::Vector;
 
-const double bsm::pi = 3.14159265358979323846;
+const float bsm::pi = 3.14159265358979323846;
 
 LorentzVector &bsm::operator +=(LorentzVector &v1, const LorentzVector &v2)
 {
@@ -27,7 +27,7 @@ LorentzVector &bsm::operator +=(LorentzVector &v1, const LorentzVector &v2)
 
 LorentzVector &bsm::operator -=(LorentzVector &v1, const LorentzVector &v2)
 {
-    double energy = v1.e() - v2.e();
+    float energy = v1.e() - v2.e();
     if (0 >= energy)
         v1.set_e(0);
     else
@@ -133,7 +133,7 @@ bool bsm::operator !=(const Vector &v1, const Vector &v2)
     return !(v1 == v2);
 }
 
-double bsm::operator *(const LorentzVector &v1, const LorentzVector &v2)
+float bsm::operator *(const LorentzVector &v1, const LorentzVector &v2)
 {
     return v1.e() * v2.e()
         - v1.px() * v2.px()
@@ -141,9 +141,9 @@ double bsm::operator *(const LorentzVector &v1, const LorentzVector &v2)
         - v1.pz() * v2.pz();
 }
 
-double bsm::dphi(const LorentzVector &v1, const LorentzVector &v2)
+float bsm::dphi(const LorentzVector &v1, const LorentzVector &v2)
 {
-    double result = phi(v1) - phi(v2);
+    float result = phi(v1) - phi(v2);
 
     while(result >= pi)
         result -= pi;
@@ -154,24 +154,24 @@ double bsm::dphi(const LorentzVector &v1, const LorentzVector &v2)
     return result;
 }
 
-double bsm::dr(const LorentzVector &v1, const LorentzVector &v2)
+float bsm::dr(const LorentzVector &v1, const LorentzVector &v2)
 {
-    double delta_phi = dphi(v1, v2);
-    double eta1 = eta(v1);
-    double eta2 = eta(v2);
+    float delta_phi = dphi(v1, v2);
+    float eta1 = eta(v1);
+    float eta2 = eta(v2);
 
     if (eta1 == -eta2
             && DBL_MAX == fabs(eta1))
         return DBL_MAX;
 
-    double delta_eta = eta1 - eta2;
+    float delta_eta = eta1 - eta2;
 
     return sqrt(delta_eta * delta_eta + delta_phi * delta_phi);
 }
 
-double bsm::et(const LorentzVector &v)
+float bsm::et(const LorentzVector &v)
 {
-    double p = momentum(v);
+    float p = momentum(v);
 
     if (0 == p
             || 0 == v.pz())
@@ -184,7 +184,7 @@ double bsm::et(const LorentzVector &v)
     }
 }
 
-double bsm::eta(const LorentzVector &v)
+float bsm::eta(const LorentzVector &v)
 {
     if (0 == v.px()
             && 0 == v.py())
@@ -198,29 +198,29 @@ double bsm::eta(const LorentzVector &v)
     }
     else
     {
-        double cosine_theta = v.pz() / momentum(v);
+        float cosine_theta = v.pz() / momentum(v);
 
         return -0.5 * log( ( 1 - cosine_theta) / (1 + cosine_theta) );
     }
 }
 
-double bsm::mass(const LorentzVector &v)
+float bsm::mass(const LorentzVector &v)
 {
-    double magnitude = v * v;
+    float magnitude = v * v;
 
     return (0 >= magnitude)
         ? 0
         : sqrt(magnitude);
 }
 
-double bsm::momentum(const LorentzVector &v)
+float bsm::momentum(const LorentzVector &v)
 {
     return sqrt(v.px() * v.px()
             + v.py() * v.py()
             + v.pz() * v.pz());
 }
 
-double bsm::phi(const LorentzVector &v)
+float bsm::phi(const LorentzVector &v)
 {
     if (v.px())
         return atan2(v.py(), v.px());
@@ -233,7 +233,7 @@ double bsm::phi(const LorentzVector &v)
         return -pi / 2;
 }
 
-double bsm::pt(const LorentzVector &v)
+float bsm::pt(const LorentzVector &v)
 {
     return sqrt(v.px() * v.px() + v.py() * v.py());
 }
